# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organization": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "all organization",
                "description": "Organization",
                "operationId": "index_organization",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "join"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Post a new data",
                "description": "Post new organization",
                "operationId": "73c0cfc20da03c7ad209a12d3578d4c4",
                "requestBody": {
                    "description": "Organization",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "description",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "john"
                                    },
                                    "image": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "bb"
                                    },
                                    "description": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "lorem"
                                    },
                                    "category_id": {
                                        "type": "number",
                                        "format": "text , exists,categories,id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/organization/{organization}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "organization",
                "description": "Returns project data",
                "operationId": "0052d6bfb12e63a53032b555112e4454",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Organization"
                ],
                "summary": "Post a new data",
                "description": "update  category  data",
                "operationId": "5d9f7f64e12ba9b0e224077903c484a5",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Id of category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass category   credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "description",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "integer",
                                        "format": "foreign",
                                        "example": "organization"
                                    },
                                    "image": {
                                        "type": "integer",
                                        "format": "foreign",
                                        "example": "cc"
                                    },
                                    "description": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "big"
                                    },
                                    "category_id": {
                                        "type": "number",
                                        "format": "foreign , exists,categories,id",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Organization"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "4bad2e386b7ed74be75470877b645c2b",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "api"
        }
    }
}